all routes protected

+++ backend/src/main/java/me/kendler/yanik/endpoints/SceneResource.java
@@ -3,9 +3,6 @@ package me.kendler.yanik.endpoints;
 import jakarta.inject.Inject;
 import me.kendler.yanik.dto.scene.*;
 import me.kendler.yanik.dto.scene.attributeDefinitions.SceneAttributeDefinitionBaseDTO;
-import me.kendler.yanik.dto.shot.ShotDTO;
-import me.kendler.yanik.dto.shot.ShotEditDTO;
-import me.kendler.yanik.model.Shotlist;
 import me.kendler.yanik.model.scene.Scene;
 import me.kendler.yanik.model.scene.attributeDefinitions.SceneAttributeDefinitionBase;
 import me.kendler.yanik.model.scene.attributeDefinitions.SceneSelectAttributeOptionDefinition;
@@ -37,24 +34,28 @@ public class SceneResource {
     @Query
     public List<SceneDTO> getScenes(UUID shotlistId) {
         userRepository.checkUserAccessRights(shotlistId, jwt);
+
         return sceneRepository.list("shotlist.id", shotlistId).stream().map(Scene::toDTO).toList();
     }
 
     @Mutation
     public SceneDTO createScene(UUID shotlistId) {
         userRepository.checkUserAccessRights(shotlistId, jwt);
+
         return sceneRepository.create(shotlistId).toDTO();
     }
 
     @Mutation
     public SceneDTO deleteScene(UUID id) {
         userRepository.checkUserAccessRights(sceneRepository.findById(id).shotlist, jwt);
+
         return sceneRepository.delete(id).toDTO();
     }
 
     @Mutation
     public SceneDTO updateScene(SceneEditDTO editDTO) {
         userRepository.checkUserAccessRights(sceneRepository.findById(editDTO.id()).shotlist, jwt);
+
         return sceneRepository.update(editDTO).toDTO();
     }
 
@@ -68,25 +69,27 @@ public class SceneResource {
     @Query
     public List<SceneAttributeDefinitionBaseDTO> getSceneAttributeDefinitions(UUID shotlistId){
         userRepository.checkUserAccessRights(shotlistId, jwt);
+
         return sceneAttributeDefinitionRepository.getAll(shotlistId);
     }
 
     @Mutation
     public SceneAttributeDefinitionBaseDTO createSceneAttributeDefinition(SceneAttributeDefinitionCreateDTO createDTO){
         userRepository.checkUserAccessRights(createDTO.shotlistId(), jwt);
+
         return sceneAttributeDefinitionRepository.create(createDTO).toDTO();
     }
 
     @Mutation
     public SceneAttributeDefinitionBase deleteSceneAttributeDefinition(Long id){
-        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getByDefinitionId(id), jwt);
+        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getShotlistByDefinitionId(id), jwt);
 
         return sceneAttributeDefinitionRepository.delete(id);
     }
 
     @Mutation
     public SceneAttributeDefinitionBase updateSceneAttributeDefinition(SceneAttributeDefinitionEditDTO editDTO) {
-        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getByDefinitionId(editDTO.id()), jwt);
+        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getShotlistByDefinitionId(editDTO.id()), jwt);
 
         return sceneAttributeDefinitionRepository.update(editDTO);
     }
@@ -100,6 +103,9 @@ public class SceneResource {
 
     @Mutation
     public SceneAttributeBase updateSceneAttribute(SceneAttributeEditDTO editDTO) {
+        SceneAttributeDefinitionBase sceneAttributeDefinition = sceneAttributeRepository.findById(editDTO.id()).definition;
+        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getShotlistByDefinitionId(sceneAttributeDefinition.id), jwt);
+
         return sceneAttributeRepository.update(editDTO);
     }
 
@@ -112,26 +118,38 @@ public class SceneResource {
 
     @Query
     public List<SceneSelectAttributeOptionDefinition> getSceneSelectAttributeOptions(Long attributeDefinitionId) {
+        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getShotlistByDefinitionId(attributeDefinitionId), jwt);
+
         return sceneSelectAttributeOptionDefinitionRepository.list("sceneSelectAttributeDefinition.id order by name", attributeDefinitionId);
     }
 
     @Query
     public List<SceneSelectAttributeOptionDefinition> searchSceneSelectAttributeOptions(SceneSelectAttributeOptionSearchDTO searchDTO){
+        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getShotlistByDefinitionId(searchDTO.sceneAttributeDefinitionId()), jwt);
+
         return sceneSelectAttributeOptionDefinitionRepository.search(searchDTO);
     }
 
     @Mutation
     public SceneSelectAttributeOptionDefinition createSceneSelectAttributeOption(SceneSelectAttributeOptionCreateDTO createDTO){
+        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getShotlistByDefinitionId(createDTO.attributeDefinitionId()), jwt);
+
         return sceneSelectAttributeOptionDefinitionRepository.create(createDTO);
     }
 
     @Mutation
     public SceneSelectAttributeOptionDefinition deleteSceneSelectAttributeOption(Long id){
+        SceneAttributeDefinitionBase sceneAttributeDefinition = sceneSelectAttributeOptionDefinitionRepository.findById(id).sceneAttributeDefinition;
+        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getShotlistByDefinitionId(sceneAttributeDefinition.id), jwt);
+
         return sceneSelectAttributeOptionDefinitionRepository.delete(id);
     }
 
     @Mutation
     public SceneSelectAttributeOptionDefinition updateSceneSelectAttributeOption(SceneSelectAttributeOptionEditDTO editDTO) {
+        SceneAttributeDefinitionBase sceneAttributeDefinition = sceneSelectAttributeOptionDefinitionRepository.findById(editDTO.id()).sceneAttributeDefinition;
+        userRepository.checkUserAccessRights(sceneAttributeDefinitionRepository.getShotlistByDefinitionId(sceneAttributeDefinition.id), jwt);
+
         return sceneSelectAttributeOptionDefinitionRepository.update(editDTO);
     }
 }
